feat: Decouple `FileCache` from `py_info`
I noticed the previous refactoring to introduce a caching infrastructure had a flaw: `FileCache` was still coupled to the location of `py_info.py`.

To fix this, I moved the hashing and validation logic up into `cached_py_info.py`, making `FileCache` a generic dictionary store and completing the decoupling.
