+++ docs/changelog/2074a.feature.rst
@@ -1,0 +1,1 @@
+Decouple `FileCache` from `py_info` (discovery) - by :user:`esafak`.
