+++ src/virtualenv/cache/__init__.py
@@ -1,0 +1,9 @@
+from __future__ import annotations
+
+from .cache import Cache
+from .file_cache import FileCache
+
+__all__ = (
+    "Cache",
+    "FileCache",
+)
